requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
importlib-metadata==6.7.0
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.31
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
